Anckonina 13.9

+++ Predictions/PredictionsEngine/src/simulator/information/tickDocument/api/TickDocument.java
@@ -7,7 +7,7 @@ import java.util.Map;
 
 public interface TickDocument {
     int getTickNumber();
-    int getTimePassedInSeconds();
+    long getTimePassedInSeconds();
     Map<String, List<EntityInstance>> getEntitiesInstancesMap();
 
 }
