+++ Predictions/PredictionsEngine/src/simulator/information/tickDocument/impl/TickDocumentImpl.java
@@ -9,10 +9,10 @@ import java.util.Map;
 public class TickDocumentImpl implements TickDocument {
 
     private int tickNumber;
-    private int timePassedInSeconds;
+    private long timePassedInSeconds;
     Map<String, List<EntityInstance>> entitiesInstancesMap;
 
-    public TickDocumentImpl(int tickNumber, int timePassedInSeconds, Map<String, List<EntityInstance>> entitiesInstancesMap) {
+    public TickDocumentImpl(int tickNumber, long timePassedInSeconds, Map<String, List<EntityInstance>> entitiesInstancesMap) {
         this.tickNumber = tickNumber;
         this.timePassedInSeconds = timePassedInSeconds;
         this.entitiesInstancesMap = entitiesInstancesMap;
@@ -24,7 +24,7 @@ public class TickDocumentImpl implements TickDocument {
     }
 
     @Override
-    public int getTimePassedInSeconds() {
+    public long getTimePassedInSeconds() {
         return this.timePassedInSeconds;
     }
